requirements.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
coverage==5.4
cryptography==3.4.6
Django==3.1.7
django-debug-toolbar==3.2
django-js-asset==1.2.2
django-mptt==0.11.0
idna==2.10
isort==5.7.0
mypy-extensions==0.4.3
pathspec==0.8.1
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
Pillow==8.1.0
pycparser==2.20
<<<<<<< Updated upstream
pyOpenSSL==20.0.1
pytz==2021.1
regex==2020.11.13
=======
>>>>>>> Stashed changes
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
<<<<<<< Updated upstream
=======
zipp==3.15.0
>>>>>>> Stashed changes
